requirements.txt
# pip install -r requirements.txt

# base
Django==4.0.6
djangorestframework==3.13.1
markdown==3.3.7
django-filter==22.1
djangorestframework-yaml==2.0.0
djangorestframework-xml==2.0.0

# devel
django-extensions==3.2.0
django-debug-toolbar==3.5.0
litecli
ipython
httpie
pytest-cov
pytest-html
jq
pytest-django
faker
